etl/pyproject.toml
[project]
name = "deep-pull-etl"
version = "0.1.0"
description = "NFL data ETL pipeline for gaming platform"
requires-python = ">=3.11"
dependencies = [
    "nfl-data-py==0.3.2",
    "pandas==2.1.4",
    "sqlalchemy==2.0.23", 
    "psycopg2-binary==2.9.9",
    "python-dotenv==1.0.0",
    "pyarrow==14.0.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
include = [
    "mvp_pipeline.py",
    "run_etl.py",
    "test_connection.py",
    "validate_data.py"
]
